refactor: separate the user field types from the types file
