+++ packages/api-postgres/src/user/UserService.ts
@@ -2,7 +2,7 @@ import bcrypt from 'bcryptjs';
 import { v4 } from 'uuid';
 import User from '../entity/User';
 import { getRepository } from 'typeorm';
-import { UserFields, UpdatableUserFields } from '../common/types/types';
+import { UserFields, UpdatableUserFields } from '../common/types/user';
 import { UserRole } from '../common/types/enums';
 import DateUtils from '../common/DateUtils';
